+++ tests/v2/08_full_workflow/test_full_workflow_local_v2.py
@@ -1,9 +1,9 @@
 import pytest
-from common import failing_workflow_UnknownError
-from common import full_workflow
-from common import full_workflow_TaskExecutionError
-from common import non_executable_task_command
-from common import non_python_task
+from common_functions import failing_workflow_UnknownError
+from common_functions import full_workflow
+from common_functions import full_workflow_TaskExecutionError
+from common_functions import non_executable_task_command
+from common_functions import non_python_task
 
 FRACTAL_RUNNER_BACKEND = "local"
